Remove PatchedGithubIntegration() class
the implementation of github.GithubIntegration from PyGithub==1.58.0
changes the signature a bit (doesn't affect us), but most importantly
internally it uses a different mechanism to raise exceptions which
passes the headers value to the exception class.

The issue that we've been working around is no longer a problem so we
don't need this class anymore.