Filter social auths only for 'github-app' instead of using .first()
Fixes KIWI-TCMS-P5, https://kiwitcms.sentry.io/issues/5501934208/?project=277775

To reproduce the original bug:

1) Login via Google OAuth - uid is recorded as the email address
2) Login via GitHhub App integration - uid is an integer
3) Try editing the GitHub App via PLUGINS -> GitHub Integration ->
   Settings

The first social_auth record will be picked up and will lead to a
ValueError b/c the query expects an integer, not a string!

With this commit the code will search for record where
provider='github-app' because this is what it needs!